Startup Version 1.0
The first version of the application. This app, if included in the qb64
folder, will run setup_osx.command if the qb64 Unix Executable does not
exist. Conversely, if the qb64 unix executable does exist, it will run
the startup sequence, and hide the Terminal window. Once the user exits
qb64, the application quits Terminal as well, then closes.